# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ESTADIA API — detention time tracking and billing for truck drivers
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OtpRequest {
  /** Brazilian phone number with country code e.g. +5511999999999 */
  telefone: string;
}

export interface OtpRequestResult {
  message: string;
}

export interface OtpVerify {
  telefone: string;
  /**
     * @minLength 4
     * @maxLength 6
     */
  codigo: string;
}

/**
 * @nullable
 */
export type MotoristaTipo = typeof MotoristaTipo[keyof typeof MotoristaTipo] | null;


export const MotoristaTipo = {
  TAC_autônomo: 'TAC autônomo',
  ETC_frota: 'ETC frota',
} as const;

export type MotoristaPlano = typeof MotoristaPlano[keyof typeof MotoristaPlano];


export const MotoristaPlano = {
  gratis: 'gratis',
  pro_mensal: 'pro_mensal',
  pro_anual: 'pro_anual',
} as const;

export interface Motorista {
  id: string;
  telefone: string;
  /** @nullable */
  nome?: string | null;
  /** @nullable */
  tipo?: MotoristaTipo;
  plano: MotoristaPlano;
  created_at: string;
}

export interface AuthSession {
  token: string;
  motorista: Motorista;
}

export type MotoristaUpdateTipo = typeof MotoristaUpdateTipo[keyof typeof MotoristaUpdateTipo];


export const MotoristaUpdateTipo = {
  TAC_autônomo: 'TAC autônomo',
  ETC_frota: 'ETC frota',
} as const;

export interface MotoristaUpdate {
  nome?: string;
  tipo?: MotoristaUpdateTipo;
}

export interface UsoMes {
  cobrancas_geradas: number;
  /**
     * null means unlimited (PRO)
     * @nullable
     */
  limite: number | null;
  plano: string;
}

export interface Veiculo {
  id: string;
  motorista_id: string;
  placa: string;
  capacidade_ton: number;
  tipo: string;
  is_padrao: boolean;
  created_at: string;
}

export interface VeiculoInput {
  placa: string;
  capacidade_ton: number;
  tipo: string;
}

export type EsperaStatus = typeof EsperaStatus[keyof typeof EsperaStatus];


export const EsperaStatus = {
  aguardando: 'aguardando',
  encerrada: 'encerrada',
  cobranca_gerada: 'cobranca_gerada',
} as const;

export interface Espera {
  id: string;
  motorista_id: string;
  veiculo_id: string;
  /** Server-recorded arrival timestamp (UTC) */
  chegada_ts: string;
  /**
     * Device-reported arrival timestamp
     * @nullable
     */
  chegada_device_ts?: string | null;
  /** @nullable */
  chegada_lat?: number | null;
  /** @nullable */
  chegada_lng?: number | null;
  /** @nullable */
  chegada_precisao_m?: number | null;
  /** @nullable */
  chegada_endereco?: string | null;
  /** @nullable */
  saida_ts?: string | null;
  status: EsperaStatus;
  /** @nullable */
  embarcador_nome?: string | null;
  /** @nullable */
  embarcador_cnpj?: string | null;
  /** @nullable */
  local_descricao?: string | null;
  fotos?: string[];
  /** Snapshot of rate at time of record */
  tarifa_ton_hora: number;
  /** @nullable */
  valor_calculado?: number | null;
  veiculo?: Veiculo;
  created_at: string;
}

export interface EsperaInput {
  veiculo_id: string;
  chegada_device_ts: string;
  chegada_lat: number;
  chegada_lng: number;
  chegada_precisao_m?: number;
  chegada_endereco?: string;
  embarcador_nome?: string;
  embarcador_cnpj?: string;
  local_descricao?: string;
}

export interface EncerrarEsperaInput {
  saida_ts: string;
}

export interface EsperaList {
  items: Espera[];
  total: number;
}

export interface EsperasResumo {
  /** Sum of paid charges */
  total_recuperado: number;
  total_esperas: number;
  total_pagas: number;
  /** Current active waiting record, if any */
  espera_ativa: Espera | null;
}

export interface FotoInput {
  /** Base64 encoded image data */
  foto_base64: string;
  /** e.g. image/jpeg */
  mime_type: string;
}

export interface FotoResult {
  url: string;
  timestamp: string;
}

export type CobrancaStatusPagamento = typeof CobrancaStatusPagamento[keyof typeof CobrancaStatusPagamento];


export const CobrancaStatusPagamento = {
  pendente: 'pendente',
  pago: 'pago',
} as const;

export interface Cobranca {
  id: string;
  espera_id: string;
  /** @nullable */
  pdf_url?: string | null;
  token_verificacao: string;
  url_verificacao?: string;
  valor: number;
  status_pagamento: CobrancaStatusPagamento;
  /** @nullable */
  enviada_via?: string | null;
  espera?: Espera;
  created_at: string;
}

export interface CobrancaInput {
  embarcador_nome?: string;
  embarcador_cnpj?: string;
}

export interface PaywallError {
  error: string;
  valor_em_jogo: number;
  plano_atual: string;
}

export interface VerificacaoPublica {
  cobranca_id: string;
  motorista_nome: string;
  chegada_ts: string;
  /** @nullable */
  saida_ts?: string | null;
  local: string;
  capacidade_ton: number;
  tarifa: number;
  valor: number;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  registrado_em: string;
}

export type AssinaturaPlano = typeof AssinaturaPlano[keyof typeof AssinaturaPlano];


export const AssinaturaPlano = {
  gratis: 'gratis',
  pro_mensal: 'pro_mensal',
  pro_anual: 'pro_anual',
} as const;

export type AssinaturaStatus = typeof AssinaturaStatus[keyof typeof AssinaturaStatus];


export const AssinaturaStatus = {
  ativo: 'ativo',
  cancelado: 'cancelado',
  expirado: 'expirado',
} as const;

/**
 * @nullable
 */
export type AssinaturaMetodo = typeof AssinaturaMetodo[keyof typeof AssinaturaMetodo] | null;


export const AssinaturaMetodo = {
  pix: 'pix',
  cartao: 'cartao',
} as const;

export interface Assinatura {
  id: string;
  motorista_id: string;
  plano: AssinaturaPlano;
  status: AssinaturaStatus;
  /** @nullable */
  expira_em?: string | null;
  /** @nullable */
  metodo?: AssinaturaMetodo;
  created_at: string;
}

export type CheckoutInputPlano = typeof CheckoutInputPlano[keyof typeof CheckoutInputPlano];


export const CheckoutInputPlano = {
  pro_mensal: 'pro_mensal',
  pro_anual: 'pro_anual',
} as const;

export interface CheckoutInput {
  plano: CheckoutInputPlano;
}

export interface CheckoutResult {
  billing_id: string;
  /** Base64 QR code image */
  pix_qr_code: string;
  /** PIX copy-paste string */
  pix_copia_cola: string;
  valor: number;
  expira_em: string;
}

export type AbacatePayWebhookPayloadData = {
  id?: string;
  status?: string;
};

export interface AbacatePayWebhookPayload {
  event: string;
  data: AbacatePayWebhookPayloadData;
}

export interface Tarifa {
  id: string;
  valor_ton_hora: number;
  vigente_desde: string;
}

export type ListEsperasParams = {
status?: ListEsperasStatus;
limit?: number;
offset?: number;
};

export type ListEsperasStatus = typeof ListEsperasStatus[keyof typeof ListEsperasStatus];


export const ListEsperasStatus = {
  aguardando: 'aguardando',
  encerrada: 'encerrada',
  cobranca_gerada: 'cobranca_gerada',
} as const;