# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ESTADIA API — detention time tracking and billing for truck drivers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP for phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "400":
          description: Invalid phone number
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and return session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid or expired OTP
...
  /esperas/{id}/cobranca:
    post:
      operationId: gerarCobranca
      tags: [cobrancas]
      summary: Generate billing document for a waiting record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CobrancaInput"
      responses:
        "201":
          description: Generated billing document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cobranca"
        "402":
          description: Paywall — plan limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaywallError"
...
  /cobrancas/{id}:
    get:
      operationId: getCobranca
      tags: [cobrancas]
      summary: Get a billing document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Billing document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cobranca"
  /cobrancas/{id}/marcar-pago:
    patch:
      operationId: marcarCobrancaPaga
      tags: [cobrancas]
      summary: Mark a billing document as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cobranca"

  # ─── PUBLIC VERIFICATION ─────────────────────────────────────────────────
...
  /assinatura/checkout:
    post:
      operationId: criarCheckout
      tags: [subscriptions]
      summary: Create AbacatePay PIX checkout for subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout created with PIX data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
  /webhooks/abacatepay:
    post:
      operationId: abacatePayWebhook
      tags: [subscriptions]
      summary: AbacatePay payment webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbacatePayWebhookPayload"
      responses:
        "200":
          description: Processed

  # ─── TARIFAS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    OtpRequest:
      type: object
      required: [telefone]
      properties:
        telefone:
          type: string
          description: Brazilian phone number with country code e.g. +5511999999999
    OtpRequestResult:
      type: object
      required: [message]
      properties:
        message:
          type: string
    OtpVerify:
      type: object
      required: [telefone, codigo]
      properties:
        telefone:
          type: string
        codigo:
          type: string
          minLength: 4
          maxLength: 6
    AuthSession:
      type: object
      required: [token, motorista]
      properties:
        token:
          type: string
        motorista:
          $ref: "#/components/schemas/Motorista"

    # ─── PROFILE ────────────────────────────────────────────────────────────
    Motorista:
      type: object
      required: [id, telefone, plano, created_at]
      properties:
        id:
          type: string
        telefone:
          type: string
        nome:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
          enum: ["TAC autônomo", "ETC frota", null]
        plano:
          type: string
          enum: [gratis, pro_mensal, pro_anual]
        created_at:
          type: string
          format: date-time
...
    Veiculo:
      type: object
      required: [id, motorista_id, placa, capacidade_ton, tipo, is_padrao, created_at]
      properties:
        id:
          type: string
        motorista_id:
          type: string
        placa:
          type: string
        capacidade_ton:
          type: number
        tipo:
          type: string
        is_padrao:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    Espera:
      type: object
      required: [id, motorista_id, veiculo_id, chegada_ts, status, tarifa_ton_hora, created_at]
      properties:
        id:
          type: string
        motorista_id:
          type: string
        veiculo_id:
          type: string
        chegada_ts:
          type: string
          format: date-time
          description: Server-recorded arrival timestamp (UTC)
        chegada_device_ts:
          type: ["string", "null"]
          format: date-time
          description: Device-reported arrival timestamp
        chegada_lat:
          type: ["number", "null"]
        chegada_lng:
          type: ["number", "null"]
        chegada_precisao_m:
          type: ["number", "null"]
        chegada_endereco:
          type: ["string", "null"]
        saida_ts:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [aguardando, encerrada, cobranca_gerada]
        embarcador_nome:
          type: ["string", "null"]
        embarcador_cnpj:
          type: ["string", "null"]
        local_descricao:
          type: ["string", "null"]
        fotos:
          type: array
          items:
            type: string
        tarifa_ton_hora:
          type: number
          description: Snapshot of rate at time of record
        valor_calculado:
          type: ["number", "null"]
        veiculo:
          $ref: "#/components/schemas/Veiculo"
        created_at:
          type: string
          format: date-time
...
    Cobranca:
      type: object
      required: [id, espera_id, token_verificacao, valor, status_pagamento, created_at]
      properties:
        id:
          type: string
        espera_id:
          type: string
        pdf_url:
          type: ["string", "null"]
        token_verificacao:
          type: string
        url_verificacao:
          type: string
        valor:
          type: number
        status_pagamento:
          type: string
          enum: [pendente, pago]
        enviada_via:
          type: ["string", "null"]
        espera:
          $ref: "#/components/schemas/Espera"
        created_at:
          type: string
          format: date-time
    CobrancaInput:
      type: object
      properties:
        embarcador_nome:
          type: string
        embarcador_cnpj:
          type: string
    PaywallError:
      type: object
      required: [error, valor_em_jogo, plano_atual]
      properties:
        error:
          type: string
        valor_em_jogo:
          type: number
        plano_atual:
          type: string
...
    CheckoutInput:
      type: object
      required: [plano]
      properties:
        plano:
          type: string
          enum: [pro_mensal, pro_anual]
    CheckoutResult:
      type: object
      required: [billing_id, pix_qr_code, pix_copia_cola, valor, expira_em]
      properties:
        billing_id:
          type: string
        pix_qr_code:
          type: string
          description: Base64 QR code image
        pix_copia_cola:
          type: string
          description: PIX copy-paste string
        valor:
          type: number
        expira_em:
          type: string
          format: date-time
    AbacatePayWebhookPayload:
      type: object
      required: [event, data]
      properties:
        event:
          type: string
        data:
          type: object
          properties:
            id:
              type: string
            status:
              type: string

    # ─── TARIFAS ────────────────────────────────────────────────────────────
...