# lib/api-zod/src/generated/types/cobranca.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ESTADIA API — detention time tracking and billing for truck drivers
 * OpenAPI spec version: 0.1.0
 */
import type { CobrancaStatusPagamento } from './cobrancaStatusPagamento';
import type { Espera } from './espera';

export interface Cobranca {
  id: string;
  espera_id: string;
  /** @nullable */
  pdf_url?: string | null;
  token_verificacao: string;
  url_verificacao?: string;
  valor: number;
  status_pagamento: CobrancaStatusPagamento;
  /** @nullable */
  enviada_via?: string | null;
  espera?: Espera;
  created_at: Date;
}