# lib/api-zod/src/generated/types/otpRequest.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * ESTADIA API — detention time tracking and billing for truck drivers
 * OpenAPI spec version: 0.1.0
 */

export interface OtpRequest {
  /** Brazilian phone number with country code e.g. +5511999999999 */
  telefone: string;
}